Rename session serializer file
- Move SessionSerialize out of session/mixin
- Update requires to use session_serializer
