+++ lib/iruby/session_serializer.rb
